Implementar registro de auditoria en ventana Modalidad Laboral

+++ servidor/javascript/controlador/catalogos/catModalidadLaboralControlador.js
@@ -14,6 +14,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.modalidaLaboralControlador = void 0;
 const accesoCarpetas_1 = require("../../libs/accesoCarpetas");
+const auditoriaControlador_1 = __importDefault(require("../auditoria/auditoriaControlador"));
 const fs_1 = __importDefault(require("fs"));
 const path_1 = __importDefault(require("path"));
 const database_1 = __importDefault(require("../../database"));
@@ -42,16 +43,30 @@ class ModalidaLaboralControlador {
     CrearMadalidadLaboral(req, res) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
-                const { modalidad } = req.body;
+                const { modalidad, user_name, ip } = req.body;
                 var VERIFICAR_MODALIDAD = yield database_1.default.query(`
                 SELECT * FROM e_cat_modalidad_trabajo WHERE UPPER(descripcion) = $1
                 `, [modalidad.toUpperCase()]);
                 if (VERIFICAR_MODALIDAD.rows[0] == undefined || VERIFICAR_MODALIDAD.rows[0] == '') {
                     const modali = modalidad.charAt(0).toUpperCase() + modalidad.slice(1).toLowerCase();
+                    // INICIAR TRANSACCION
+                    yield database_1.default.query('BEGIN');
                     const response = yield database_1.default.query(`
                     INSERT INTO e_cat_modalidad_trabajo (descripcion) VALUES ($1) RETURNING *
                     `, [modali]);
                     const [modalidadLaboral] = response.rows;
+                    // REGISTRAR AUDITORIA
+                    yield auditoriaControlador_1.default.InsertarAuditoria({
+                        tabla: 'e_cat_modalidad_trabajo',
+                        usuario: user_name,
+                        accion: 'I',
+                        datosOriginales: '',
+                        datosNuevos: JSON.stringify(modalidadLaboral),
+                        ip,
+                        observacion: null
+                    });
+                    // FINALIZAR TRANSACCION
+                    yield database_1.default.query('COMMIT');
                     if (modalidadLaboral) {
                         return res.status(200).jsonp({ message: 'Registro guardado.', status: '200' });
                     }
@@ -64,6 +79,8 @@ class ModalidaLaboralControlador {
                 }
             }
             catch (error) {
+                // ROLLBACK SI HAY ERROR
+                yield database_1.default.query('ROLLBACK');
                 return res.status(500).jsonp({ message: 'error', status: '500' });
             }
         });
@@ -72,7 +89,7 @@ class ModalidaLaboralControlador {
     EditarModalidadLaboral(req, res) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
-                const { id, modalidad } = req.body;
+                const { id, modalidad, user_name, ip } = req.body;
                 const modali = modalidad.charAt(0).toUpperCase() + modalidad.slice(1).toLowerCase();
                 const modalExiste = yield database_1.default.query(`
                 SELECT * FROM e_cat_modalidad_trabajo WHERE UPPER(descripcion) = $1
@@ -81,11 +98,25 @@ class ModalidaLaboralControlador {
                     return res.status(200).jsonp({ message: 'Modalidad Laboral ya esiste en el sistema.', status: '300' });
                 }
                 else {
+                    // INICIAR TRANSACCION
+                    yield database_1.default.query('BEGIN');
                     const response = yield database_1.default.query(`
                     UPDATE e_cat_modalidad_trabajo SET descripcion = $2
                     WHERE id = $1 RETURNING *
                     `, [id, modali]);
                     const [modalidadLaboral] = response.rows;
+                    // REGISTRAR AUDITORIA
+                    yield auditoriaControlador_1.default.InsertarAuditoria({
+                        tabla: 'e_cat_modalidad_trabajo',
+                        usuario: user_name,
+                        accion: 'U',
+                        datosOriginales: JSON.stringify(modalExiste.rows),
+                        datosNuevos: JSON.stringify(modalidadLaboral),
+                        ip,
+                        observacion: null
+                    });
+                    // FINALIZAR TRANSACCION
+                    yield database_1.default.query('COMMIT');
                     if (modalidadLaboral) {
                         return res.status(200).jsonp({ message: 'Registro actualizado.', status: '200' });
                     }
@@ -95,6 +126,8 @@ class ModalidaLaboralControlador {
                 }
             }
             catch (error) {
+                // ROLLBACK SI HAY ERROR
+                yield database_1.default.query('ROLLBACK');
                 return res.status(500).jsonp({ message: 'error', status: '500' });
             }
         });
@@ -104,12 +137,48 @@ class ModalidaLaboralControlador {
         return __awaiter(this, void 0, void 0, function* () {
             try {
                 const id = req.params.id;
+                const { user_name, ip } = req.body;
+                // INICIAR TRANSACCION
+                yield database_1.default.query('BEGIN');
+                // CONSULTAR DATOS ANTES DE ELIMINAR
+                const modalidad = yield database_1.default.query(`
+                SELECT * FROM e_cat_modalidad_trabajo WHERE id = $1
+                `, [id]);
+                const [modalidadLaboral] = modalidad.rows;
+                if (!modalidadLaboral) {
+                    yield auditoriaControlador_1.default.InsertarAuditoria({
+                        tabla: 'e_cat_modalidad_trabajo',
+                        usuario: user_name,
+                        accion: 'D',
+                        datosOriginales: '',
+                        datosNuevos: '',
+                        ip,
+                        observacion: `Error al eliminar el registro con id: ${id}, no se encuentra el registro.`
+                    });
+                    // FINALIZAR TRANSACCION
+                    yield database_1.default.query('COMMIT');
+                    return res.status(404).jsonp({ message: 'No se encuentra el registro.' });
+                }
                 yield database_1.default.query(`
                 DELETE FROM e_cat_modalidad_trabajo WHERE id = $1
                 `, [id]);
+                // REGISTRAR AUDITORIA
+                yield auditoriaControlador_1.default.InsertarAuditoria({
+                    tabla: 'e_cat_modalidad_trabajo',
+                    usuario: user_name,
+                    accion: 'D',
+                    datosOriginales: JSON.stringify(modalidadLaboral),
+                    datosNuevos: '',
+                    ip,
+                    observacion: null
+                });
+                // FINALIZAR TRANSACCION
+                yield database_1.default.query('COMMIT');
                 res.jsonp({ message: 'Registro eliminado.' });
             }
             catch (error) {
+                // ROLLBACK SI HAY ERROR
+                yield database_1.default.query('ROLLBACK');
                 return res.jsonp({ message: 'error' });
             }
         });
@@ -232,19 +301,32 @@ class ModalidaLaboralControlador {
     CargarPlantilla(req, res) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
-                const plantilla = req.body;
-                console.log('datos Modalidad laboral: ', plantilla);
+                const { plantilla, user_name, ip } = req.body;
                 var contador = 1;
                 var respuesta;
                 plantilla.forEach((data) => __awaiter(this, void 0, void 0, function* () {
                     // DATOS QUE SE GUARDARAN DE LA PLANTILLA INGRESADA
                     const { item, modalida_laboral, observacion } = data;
                     const modalidad = modalida_laboral.charAt(0).toUpperCase() + modalida_laboral.slice(1).toLowerCase();
+                    // INICIO DE TRANSACCION
+                    yield database_1.default.query('BEGIN');
                     // REGISTRO DE LOS DATOS DE MODLAIDAD LABORAL
                     const response = yield database_1.default.query(`
                     INSERT INTO e_cat_modalidad_trabajo (descripcion) VALUES ($1) RETURNING *
                     `, [modalidad]);
                     const [modalidad_la] = response.rows;
+                    // REGISTRAR AUDITORIA
+                    yield auditoriaControlador_1.default.InsertarAuditoria({
+                        tabla: 'e_cat_modalidad_trabajo',
+                        usuario: user_name,
+                        accion: 'I',
+                        datosOriginales: '',
+                        datosNuevos: JSON.stringify(modalidad_la),
+                        ip,
+                        observacion: null
+                    });
+                    // FINALIZAR TRANSACCION
+                    yield database_1.default.query('COMMIT');
                     if (contador === plantilla.length) {
                         if (modalidad_la) {
                             return respuesta = res.status(200).jsonp({ message: 'ok' });
@@ -257,6 +339,8 @@ class ModalidaLaboralControlador {
                 }));
             }
             catch (error) {
+                // ROLLBACK SI HAY ERROR
+                yield database_1.default.query('ROLLBACK');
                 return res.status(500).jsonp({ message: error });
             }
         });
